removes react and graphql corp curriculums
