+++ src/components/curriculum/index.js
@@ -3,13 +3,11 @@ import CurriculumPartTime from './CurriculumPartTime'
 import CurriculumReactNative from './CurriculumReactNative'
 import MarketingCard from './MarketingCard'
 import CurriculumAdvancedReact from './CurriculumAdvancedReact'
-import CurriculumCorpReact from './CurriculumCorpReact'
 import CurriculumCorpRNative from './CurriculumCorpRNative'
 import CurriculumGraphQLBootcamp from './CurriculumGraphQLBootcamp'
 import CurriculumGraphQLAPI from './CurriculumGraphQLAPI'
 import FullCurriculumsReact from './FullCurriculumsReact'
 import CurriculumCorporateReact from './CurriculumCorporateReact'
-import CurriculumCorpGraphQL from './CurriculumCorpGraphQL'
 import FullCurriculum from './FullCurriculum'
 import CurriculumReactFundamentals from './CurriculumReactFundamentals'
 import CurriculumGraphQLWorkshops from './CurriculumGraphQLWorkshops'
@@ -25,9 +23,7 @@ export {
   CurriculumGraphQLAPI,
   CurriculumCorpRNative,
   FullCurriculumsReact,
-  CurriculumCorpReact,
   CurriculumCorporateReact,
-  CurriculumCorpGraphQL,
   FullCurriculum,
   CurriculumReactFundamentals,
   CurriculumGraphQLWorkshops,
