+++ src/pages/community/meetups/5d07b4fb06051b7d3bcb0d1c.md
@@ -11,7 +11,7 @@ Le workshop se compose de 2 parties:
 • Exercice pratique
 
 Nombre de places limité aux 20 premiers inscrits
- 
+
 [David Leuliette est Coach](/about-us#david-leuliette) à ReactGraphQL Academy
 
 Pas de niveau prérequis, juste savoir comment cloner un projet et créer un composant React avec `create-react-app`.
